pyproject.toml
[project]
name = "nuv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "connect6-pro-python",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
    "pycryptodome>=3.20.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
]

[tool.uv.sources]
connect6-pro-python = { git = "https://github.com/bay6-ai/connect6-pro-python.git", rev = "dev" }
